CurioApplication/requirements.txt
flask
flask_cors
requests
python-dotenv
apscheduler
pytz